# api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: Nova Mart Commerce API
  version: 1.0.0
  description: |
    Public HTTP contract for the **Nova Mart** online shopping platform.

    This document is the **source of truth**. It was authored *before* any service
    implementation, and every backend endpoint and every frontend API client method
    is derived from it. Changing behaviour means changing this file first.

    ## Architecture

    Clients never talk to a service directly. All traffic enters through the
    **API Gateway** (`/api/v1/**`), which authenticates the caller, strips any
    client-supplied identity headers, injects trusted ones, and routes to the
    owning service.

    | Path prefix                                   | Owning service       | Owns database     |
    | --------------------------------------------- | -------------------- | ----------------- |
    | `/auth`, `/users`                              | auth-service         | `auth_db`         |
    | `/products`, `/categories`, `/brands`          | product-service      | `product_db`      |
    | `/cart`                                        | cart-service         | `cart_db`         |
    | `/orders`                                      | order-service        | `order_db`        |
    | `/payments`                                    | payment-service      | `payment_db`      |
    | `/inventory`                                   | inventory-service    | `inventory_db`    |
...

...

    ## Response envelope

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      operationId: register
      summary: Register a new shopper account
      description: Creates a `USER` account and immediately issues a token pair. Email must be unique.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              firstName: Ananya
              lastName: Iyer
              email: ananya.iyer@example.com
              password: Str0ng!Pass
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthTokens'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: An account with this email already exists
                errorCode: EMAIL_ALREADY_EXISTS
                timestamp: '2026-02-14T09:31:08.442Z'
                path: /api/v1/auth/register
  /auth/login:
    post:
      tags: [Authentication]
      operationId: login
      summary: Exchange credentials for a token pair
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: demo@novamart.dev
              password: Demo@12345
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthTokens'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Bad credentials or disabled account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: Incorrect email or password
                errorCode: INVALID_CREDENTIALS
                timestamp: '2026-02-14T09:32:11.001Z'
                path: /api/v1/auth/login
  /auth/refresh:
    post:
      tags: [Authentication]
      operationId: refreshToken
      summary: Rotate an expired access token
      description: Consumes a refresh token and issues a new pair. The presented refresh token is revoked (single use).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New token pair issued
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthTokens'
        '401':
          description: Refresh token invalid, expired or already consumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags: [Authentication]
      operationId: logout
      summary: Revoke the caller's refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '204':
          description: Token revoked
        '401':
          $ref: '#/components/responses/Unauthorized'

  # =====================================================================
  #  USERS
  # =====================================================================
  /users/me:
    get:
      tags: [Users]
      operationId: getMyProfile
      summary: Get the signed-in user's profile
      responses:
        '200':
          description: Profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [Users]
      operationId: updateMyProfile
      summary: Update the signed-in user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserProfile'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/addresses:
    get:
      tags: [Users]
      operationId: listAddresses
      summary: List the caller's saved addresses
      responses:
        '200':
          description: Address book
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Address'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Users]
      operationId: createAddress
      summary: Add an address
      description: The first address saved for a user automatically becomes the default.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressRequest'
      responses:
        '201':
          description: Address created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Address'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/addresses/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      tags: [Users]
      operationId: updateAddress
      summary: Update an address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressRequest'
      responses:
        '200':
          description: Address updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Address'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Users]
      operationId: deleteAddress
      summary: Delete an address
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List all users (admin)
      description: Administrator only. Backs the admin dashboard user table.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: search
          in: query
          description: Case-insensitive match against name or email
          schema:
            type: string
      responses:
        '200':
          description: Paged users
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PagedUsers'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # =====================================================================
  #  PRODUCTS
  # =====================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT issued by auth-service. Send as `Authorization: Bearer <accessToken>`.'
  parameters:
    Page:
      name: page
      in: query
      description: Zero-based page index
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      description: Page size
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 12
  responses:
    ValidationError:
      description: The request body or query string failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Request validation failed
            errorCode: VALIDATION_FAILED
            timestamp: '2026-02-14T09:40:12.882Z'
            path: /api/v1/auth/register
            fieldErrors:
              - field: password
                message: Password must be at least 8 characters
              - field: email
                message: Must be a well-formed email address
    Unauthorized:
      description: Missing, malformed or expired access token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Authentication is required to access this resource
            errorCode: UNAUTHORIZED
            timestamp: '2026-02-14T09:41:03.114Z'
            path: /api/v1/cart
    Forbidden:
      description: Authenticated, but not allowed to perform this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: You do not have permission to perform this action
            errorCode: FORBIDDEN
            timestamp: '2026-02-14T09:42:55.900Z'
            path: /api/v1/products
...
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Product not found
            errorCode: PRODUCT_NOT_FOUND
            timestamp: '2026-02-14T09:43:30.720Z'
            path: /api/v1/products/8f14e45f-ea8f-4b2a-9d1c-3c2b7a51de90
  schemas:
...
    ApiResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Product retrieved successfully
        data:
          description: Operation-specific payload. Absent for `204` responses.
    ErrorResponse:
      type: object
      required: [success, message, errorCode, timestamp]
      properties:
        success:
          type: boolean
          enum: [false]
        message:
          type: string
          description: Human-readable, safe to show a shopper. Never contains a stack trace.
        errorCode:
          type: string
          description: Stable machine-readable identifier. Clients branch on this, not on the message text.
          example: PRODUCT_NOT_FOUND
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        fieldErrors:
          type: array
          description: Present only when `errorCode` is `VALIDATION_FAILED`.
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    PageMeta:
      type: object
      required: [page, size, totalElements, totalPages, first, last]
      properties:
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 12
        totalElements:
          type: integer
          format: int64
          example: 24
        totalPages:
          type: integer
          example: 2
        first:
          type: boolean
        last:
          type: boolean

    # ---------- auth ----------
    RegisterRequest:
      type: object
      required: [firstName, lastName, email, password]
      properties:
        firstName:
          type: string
          minLength: 1
          maxLength: 60
        lastName:
          type: string
          minLength: 1
          maxLength: 60
        email:
          type: string
          format: email
          maxLength: 180
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 100
          description: Stored only as a BCrypt hash. Never logged, never returned.
        phone:
          type: string
          maxLength: 20
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    AuthTokens:
      type: object
      required: [accessToken, refreshToken, tokenType, expiresIn, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          enum: [Bearer]
        expiresIn:
          type: integer
          description: Access token lifetime in seconds
          example: 3600
        user:
          $ref: '#/components/schemas/UserProfile'
    UserProfile:
      type: object
      required: [id, firstName, lastName, email, roles, createdAt]
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        roles:
          type: array
          items:
            type: string
            enum: [USER, ADMIN]
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UpdateProfileRequest:
      type: object
      required: [firstName, lastName]
      properties:
        firstName:
          type: string
          minLength: 1
          maxLength: 60
        lastName:
          type: string
          minLength: 1
          maxLength: 60
        phone:
          type: string
          maxLength: 20
          nullable: true
    PagedUsers:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserProfile'
        page:
          $ref: '#/components/schemas/PageMeta'
    Address:
      type: object
      required: [id, label, recipientName, phone, line1, city, state, postalCode, country, isDefault]
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
          example: Home
        recipientName:
          type: string
        phone:
          type: string
        line1:
          type: string
        line2:
          type: string
          nullable: true
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
          example: India
        isDefault:
          type: boolean
    AddressRequest:
      type: object
      required: [label, recipientName, phone, line1, city, state, postalCode, country]
      properties:
        label:
          type: string
          maxLength: 40
        recipientName:
          type: string
          maxLength: 120
        phone:
          type: string
          maxLength: 20
        line1:
          type: string
          maxLength: 200
        line2:
          type: string
          maxLength: 200
          nullable: true
        city:
          type: string
          maxLength: 80
        state:
          type: string
          maxLength: 80
        postalCode:
          type: string
          maxLength: 16
        country:
          type: string
          maxLength: 80
        isDefault:
          type: boolean
          default: false

    # ---------- catalogue ----------
...